fw-6473, move base tests for non-site APIs to own module
